requirements.txt
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
ibm-cloud-sdk-core==1.5.1
ibm-db==3.0.1
ibm-watson==4.5.0
idna==2.9
isort==4.3.21
joblib==0.15.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.5
pandas==1.0.4
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
websocket-client==0.48.0
wrapt==1.12.1
